+++ R/description.R
@@ -30,6 +30,8 @@
 #' @param fields A named list of fields to add to `DESCRIPTION`, potentially
 #'   overriding default values. See [use_description()] for how you can set
 #'   personalized defaults using package options
+#' @param stop_for_name Whether to throw an error if the package name is not
+#'   valid for CRAN
 #' @seealso The [description chapter](https://r-pkgs.org/description.html#dependencies)
 #'   of [R Packages](https://r-pkgs.org).
 #' @export
@@ -41,9 +43,9 @@
 #'
 #' use_description_defaults()
 #' }
-use_description <- function(fields = NULL) {
+use_description <- function(fields = NULL, stop_for_name = FALSE) {
   name <- project_name()
-  check_package_name(name)
+  check_package_name(name, stop_for_name = stop_for_name)
   fields <- fields %||% list()
   check_is_named_list(fields)
   fields[["Package"]] <- name
@@ -97,15 +99,20 @@ build_description_list <- function(fields = list()) {
   compact(utils::modifyList(defaults, fields))
 }
 
-check_package_name <- function(name) {
+check_package_name <- function(name, stop_for_name) {
   if (!valid_package_name(name)) {
-    ui_stop(c(
-      "{ui_value(name)} is not a valid package name. It should:",
+    msg <- c(
+      "{ui_value(name)} is not a valid package name. To be allowed on CRAN, it should:",
       "* Contain only ASCII letters, numbers, and '.'",
       "* Have at least two characters",
       "* Start with a letter",
       "* Not end with '.'"
-    ))
+    )
+    if (stop_for_name) {
+      ui_stop(msg)
+    } else {
+      ui_warn(stop_for_name)
+    }
   }
 
 }